consolidate risk adjustment
